.build_old_20260228_192831/extensions/omni-client/backend/requirements.txt
fastapi
uvicorn
websockets
# Transcriber
faster-whisper
# Agent Framework (Lightweight)
smolagents[toolkit]
# Cloud Brain (API Client)
huggingface_hub
requests
# Utilities
python-dotenv
psutil
numpy
soundfile
aiofiles
# Safe Eval
python-multipart
duckduckgo-search
ddgs
